sdl2: timer rework

+++ Engine/ac/character.cpp
@@ -80,7 +80,7 @@ extern int said_speech_line;
 extern int numscreenover;
 extern int said_text;
 extern int our_eip;
-extern int update_music_at;
+extern uint64_t update_music_at;
 extern int current_screen_resolution_multiplier;
 extern int cur_mode;
 extern CCCharacter ccDynamicCharacter;
